requirements.txt
# Core dependencies for MAIRA-2
pillow>=10.0.0
protobuf>=4.25.0
sentencepiece>=0.1.99
torch>=2.0.0
transformers>=4.48.0,<4.52
requests>=2.31.0
accelerate>=0.20.0

# Evaluation metrics dependencies
nltk>=3.8.0                    # For BLEU and METEOR metrics
rouge-score>=0.1.2             # For ROUGE-L metric
bert-score>=0.3.13             # For BERTScore metric
numpy>=1.24.0                  # For numerical operations
tqdm>=4.65.0                   # For progress bars

# Optional: Clinical metrics (may require additional setup)
# radgraph>=0.1.0              # For RGER metric (requires separate installation)
# pydicom>=2.3.0               # For loading DICOM files
